Move the th-TH unit test to Manual test project. Add isUsingManagedSNI annotation to unit test.
